registration/requirements.txt
gtts
SpeechRecognition
playsound
google-generativeai
django
django-cors-headers
pydaudio